[NCL-6194] Extract ReposiotryIdentifier constants and add Distribution archive repository type
